requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==20.1.0
tensorflow==2.15.0
tf-keras==2.15.1
flask_httpauth==4.8.0
python-dotenv==1.0.1